HONEY-SHIELD2.0/requirements.txt
flask==2.3.3
paramiko==3.3.1
scapy==2.5.0
python-telegram-bot==20.6
requests==2.31.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
gunicorn==21.2.0
colorama==0.4.6
# Additional required dependencies
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1
python-socketio==5.12.1
python-engineio==4.11.2
bidict==0.23.1
simple-websocket==1.1.0 